refactor: Moving render utils into separate components
